--- pages/type-email.js
@@ -1,33 +1,0 @@
-// import { Button } from "bootstrap";
-import { Col, Container, Row } from "react-bootstrap";
-import CButton from "../component/CButton";
-import Input from "../component/Input";
-import login from "../styles/form-login.module.scss";
-import input from "../styles/input.module.scss";
-import {MdOutlineMailOutline} from "react-icons/md"
-import {BiLockAlt,} from 'react-icons/bi'
-import LayoutLogin from "../component/LayoutLogin";
-// import NavbarComponent from "../component/NavbarComponent";
-
-const EmailFogotPassword = () =>{
-    return (
-     <LayoutLogin>
-            <div className="vh-100 overflow-auto">
-                <form className="overflow-auto p-5 me-5 mt-5">
-                    <div className="fs-1 text-primary fw-bold">Did You Forgot Your Password? Don’t Worry, You Can Reset Your Password In a Minutes.</div>
-                    <div className="fs-5 text-primary mt-5">To reset your password, you must type your e-mail and we will send a link to your email and you will be directed to the reset password screens.</div>
-                        <div className="mt-5">
-                            <div className={input.inputContainer}>
-                                <MdOutlineMailOutline/><Input type="text" className={input.textLoginSignup} placeholder="Enter your e-mail"/>
-                            </div>
-                            <div className="mt-5">
-                                <CButton className={input.button}>Confirm</CButton>
-                            </div>
-                        </div>
-                    </form>
-                </div>   
-     </LayoutLogin>   
-    )
-}
-
-export default EmailFogotPassword
